이미지 파일 계산 타입스크립트로 변경
